// src/types/sockets.events.ts
export enum SOCKET_EVENTS_ENUM {
  BOARD_JOIN = 'BOARD_JOIN',
  BOARD_LEAVE = 'BOARD_LEAVE',
  BOARD_MODIFY_VIEW_SET = 'BOARD_MODIFY_VIEW_SET',
  BOARD_MODIFY_VIEW_SET_RESULT = 'BOARD_MODIFY_VIEW_SET_RESULT',
  BOARD_MODIFY_TITLE = 'BOARD_MODIFY_TITLE',
  BOARD_MODIFY_TITLE_RESULT = 'BOARD_MODIFY_TITLE_RESULT',
  BOARD_ARCHIVE = 'BOARD_ARCHIVE',
  BOARD_ARCHIVE_RESULT = 'BOARD_ARCHIVE_RESULT',
  BOARD_LIST_ARCHIVE = 'BOARD_LIST_ARCHIVE',
  BOARD_LIST_ARCHIVE_RESULT = 'BOARD_LIST_ARCHIVE_RESULT',
  BOARD_MOVE_LIST_POSITION = 'BOARD_MOVE_LIST_POSITION',
  BOARD_MOVE_LIST_POSITION_RESULT = 'BOARD_MOVE_LIST_POSITION_RESULT',
  BOARD_CREATE_NEW_TAG = 'BOARD_CREATE_NEW_TAG',
  BOARD_CREATE_NEW_TAG_RESULT = 'BOARD_CREATE_NEW_TAG_RESULT',
  BOARD_MODIFY_TAG = 'BOARD_MODIFY_TAG',
  BOARD_MODIFY_TAG_RESULT = 'BOARD_MODIFY_TAG_RESULT',
  BOARD_DELETE_TAG = 'BOARD_DELETE_TAG',
  BOARD_DELETE_TAG_RESULT = 'BOARD_DELETE_TAG_RESULT',
  BOARD_CARD_CREATE = 'BOARD_CARD_CREATE',
  BOARD_CARD_CREATE_RESULT = 'BOARD_CARD_CREATE_RESULT',
  BOARD_CARD_MODIFY = 'BOARD_CARD_MODIFY',
  BOARD_CARD_MODIFY_RESULT = 'BOARD_CARD_MODIFY_RESULT',
  ATTACH_TAG_TO_CARD = 'ATTACH_TAG_TO_CARD',
  ATTACH_TAG_TO_CARD_RESULT = 'ATTACH_TAG_TO_CARD_RESULT',
  REMOVE_TAG_FROM_CARD = 'REMOVE_TAG_FROM_CARD',
  REMOVE_TAG_FROM_CARD_RESULT = 'REMOVE_TAG_FROM_CARD_RESULT',
  ADD_NEW_CARD_COMMENT = 'ADD_NEW_CARD_COMMENT',
  ADD_NEW_CARD_COMMENT_RESULT = 'ADD_NEW_CARD_COMMENT_RESULT',
  MODIFT_CARD_COMMENT = 'MODIFT_CARD_COMMENT',
  MODIFT_CARD_COMMENT_RESULT = 'MODIFT_CARD_COMMENT_RESULT',
  DELETE_CARD_COMMENT = 'DELETE_CARD_COMMENT',
  DELETE_CARD_COMMENT_RESULT = 'DELETE_CARD_COMMENT_RESULT',
  BOARD_CREATE_LIST = 'BOARD_CREATE_LIST',
  BOARD_CREATE_LIST_RESULT = 'BOARD_CREATE_LIST_RESULT',
  BOARD_MODIFY_LIST_TITLE = 'BOARD_MODIFY_LIST_TITLE',
  BOARD_MODIFY_LIST_TITLE_RESULT = 'BOARD_MODIFY_LIST_TITLE_RESULT',
  BOARD_ARCHIVE_LIST = 'BOARD_ARCHIVE_LIST',
  BOARD_ARCHIVE_LIST_RESULT = 'BOARD_ARCHIVE_LIST_RESULT',
  BOARD_MODIFY_MEMBER_PERMISSION = 'BOARD_MODIFY_MEMBER_PERMISSION',
  BOARD_MODIFY_MEMBER_PERMISSION_RESULT = 'BOARD_MODIFY_MEMBER_PERMISSION_RESULT',
  BOARD_DELETE_MEMBER = 'BOARD_DELETE_MEMBER',
  BOARD_DELETE_MEMBER_RESULT = 'BOARD_DELETE_MEMBER_RESULT',
  BOARD_ADD_MEMBER = 'BOARD_ADD_MEMBER',
  BOARD_ADD_MEMBER_RESULT = 'BOARD_ADD_MEMBER_RESULT',
}